fix : self주입하여 사용되고 있는 메서드를 별도 Bean으로 분리
