+++ payment/src/main/java/com/devticket/payment/wallet/domain/repository/WalletRepository.java
@@ -8,6 +8,8 @@ public interface WalletRepository {
 
     Optional<Wallet> findByUserId(UUID userId);
 
+    void insertWalletIfAbsent(UUID userId);
+
     Optional<Wallet> findByUserIdForUpdate(UUID userId);
 
     Wallet save(Wallet wallet);
